+++ gnosis_vpn-lib/src/core/runner.rs
@@ -2,15 +2,12 @@
 //! These function expect to be spawn and will deliver their result or progress via channels.
 
 use backon::{ExponentialBuilder, Retryable};
-use bytesize::ByteSize;
 use edgli::blokli::{IncentiveOperations, make_incentive_operations};
 use edgli::hopr_lib::api::node::HoprState;
 use edgli::hopr_lib::api::types::primitive::prelude::Address;
 use edgli::hopr_lib::builder::Keypair;
 use edgli::hopr_lib::exports::network::types::types::IpProtocol;
-use edgli::hopr_lib::exports::transport::SurbBalancerConfig;
 use edgli::{BlockchainConnectorConfig, EdgliInitState};
-use human_bandwidth::re::bandwidth::Bandwidth;
 use rand::prelude::*;
 use serde::Deserialize;
 use serde_json::json;
@@ -132,14 +129,6 @@ pub enum Error {
     FundingTool(String),
     #[error("IncentiveOperations creation error: {0}")]
     IncentiveOperationsCreation(String),
-}
-
-#[derive(Debug, Error)]
-pub enum SurbConfigError {
-    #[error("Response buffer byte size too small")]
-    ResponseBufferTooSmall,
-    #[error("Max SURB upstream bandwidth cannot be zero")]
-    MaxSurbUpstreamCannotBeZero,
 }
 
 #[derive(Debug, Deserialize)]
@@ -664,22 +653,3 @@ impl Display for Results {
     }
 }
 
-pub fn to_surb_balancer_config(
-    response_buffer: ByteSize,
-    max_surb_upstream: Bandwidth,
-) -> Result<SurbBalancerConfig, SurbConfigError> {
-    // Buffer worth at least 2 reply packets
-    if response_buffer.as_u64() < 2 * edgli::hopr_lib::exports::transport::SESSION_MTU as u64 {
-        return Err(SurbConfigError::ResponseBufferTooSmall);
-    }
-    if max_surb_upstream.is_zero() {
-        return Err(SurbConfigError::MaxSurbUpstreamCannotBeZero);
-    }
-    let config = SurbBalancerConfig {
-        target_surb_buffer_size: response_buffer.as_u64() / edgli::hopr_lib::exports::transport::SESSION_MTU as u64,
-        max_surbs_per_sec: (max_surb_upstream.as_bps() as usize / (8 * edgli::hopr_lib::exports::transport::SURB_SIZE))
-            as u64,
-        ..Default::default()
-    };
-    Ok(config)
-}
